added function to create document chunks
