+++ backend/.gitignore
@@ -1,3 +1,2 @@
-
-
 .env
+data
