+++ backend/app/services/llm.py
@@ -26,46 +26,3 @@ def build_context(docs):
 
 def generate_answer(query: str, chat_history: list):
     return ""
-
-# texts = load_and_split_pdf("data")
-# retriever = initialize_retriever(texts)
-
-# print("LLM Chain Initialized")
-
-# def build_context(docs):
-#     return "\n".join([doc.page_content for doc in docs])[:1500]
-
-# def generate_answer(query: str, chat_history: list):
-#     try:
-#         docs = retriever.invoke(query)
-#         context = build_context(docs)
-      
-#         formatter = RunnableLambda(lambda d: prompt.format(**d))
-#         chain = (
-#             RunnableMap({
-#                  "context": lambda x: x["context"],
-#                  "chat_history": lambda x: x["chat_history"],
-#                  "query": lambda x: x["query"]
-#                          })
-#             | formatter      # Convert dict to formatted string
-#             | LLM            # HuggingFaceEndpoint expects a string prompt
-#             | (lambda x: x.content if hasattr(x, "content") else x)
-#                 )
-
-#         inputs = {
-#             "context": context,
-#             "chat_history": chat_history,
-#             "query": query
-#         }
-
-#         print("Calling LLM chain with inputs:", inputs)
-#         result = chain.invoke(inputs)
-#         print("LLM Response:", result)
-
-#         return result 
-
-#     except Exception as e:
-#         import traceback
-#         print("ERROR in generate_answer:", repr(e))
-#         traceback.print_exc()
-#         return "Sorry, something went wrong while generating the answer."
