+++ backend/app/services/vector_store.py
@@ -1,5 +1,7 @@
 from langchain_mongodb.vectorstores import MongoDBAtlasVectorSearch
 from langchain_huggingface import HuggingFaceEmbeddings
+from langchain.schema import Document
+from typing import List
 from app.db.connect import connectDB
 
 db = connectDB()
@@ -18,3 +20,15 @@ async def create_vectorstore() -> MongoDBAtlasVectorSearch | None:
     except Exception as e:
         print(f"error in retreiver: {e}")
         return None
+    
+async def store_in_vectorstore(documents: List[Document]) -> bool:
+    try: 
+        vector_store = await create_vectorstore()
+        if vector_store is None:
+            raise Exception("Vectorstore not available")
+        doc_ids = await vector_store.aadd_documents(documents=documents)
+        print(f"{len(doc_ids)} Documents stored")
+        return True
+    except Exception as e:
+        print(f"error in retreiver: {e}")
+        return False
